src/requirements.txt
absl-py==2.1.0
astroid==3.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
CTkMenuBar==0.3
CTkTable==1.1
CTkToolTip==0.8
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
dill==0.3.8
flatbuffers==24.3.25
fonttools==4.47.2
gast==0.5.4
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.62.1
h5py==3.10.0
idna==3.6
isort==5.13.2
joblib==1.3.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mplcursors==0.5.3
namex==0.0.7
numpy==1.26.3
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
Pillow==10.1.0
platformdirs==4.2.0
protobuf==4.25.3
Pygments==2.17.2
pylint==3.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pywinstyles==1.6
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
tomlkit==0.12.4
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
